Move methods with logic from DatabaseVo to DatabaseService.
