+++ src/Vo/DatabaseVO.php
@@ -4,8 +4,6 @@
 namespace App\Vo;
 
 
-use App\Dto\EntityCollectionDTO;
-use App\Interface\EntityExtended;
 use App\Entity\Department;
 use App\Entity\Place;
 use App\Entity\Region;
@@ -60,155 +58,6 @@ class DatabaseVO
         $this->regions = $this->entityManager->getRepository(Region::class)->findAll();
         $this->departments = $this->entityManager->getRepository(Department::class)->findAll();
         $this->places = $this->entityManager->getRepository(Place::class)->findAll();
-    }
-
-    /** Return entities (Region, Department, Source, Type) related by Dances
-     *
-     * @return EntityCollectionDTO
-     */
-    public function getEntitiesRelatedByDances(): EntityCollectionDTO
-    {
-        $entityCollection = new EntityCollectionDTO();
-
-        $entityCollection->setSources($this->setDances($this->sources));
-        $entityCollection->setTypes($this->setDances($this->types));
-        $entityCollection->setRegions($this->setDances($this->regions));
-        $entityCollection->setDepartments($this->setDances($this->departments));
-
-        return $entityCollection;
-    }
-
-    /**
-     * @param array <int, EntityExtended> $entityCollection
-     *
-     * @return EntityExtended[]
-     */
-    public function setDances(array $entityCollection): array
-    {
-        foreach ($entityCollection as $entity) {
-            $versions = $this->getVersionsByEntity($entity);
-            $dances = array();
-
-            foreach ($versions as $version) {
-                $dances[] = $version->getDance();
-            }
-
-            $entity->setDances($dances);
-        }
-
-        return $entityCollection;
-    }
-
-    /** Get related versions for Region or Department or Source or Type.
-     * @param EntityExtended $entity
-     *
-     * @return array<int, Version>
-     */
-    public function getVersionsByEntity(EntityExtended $entity): array
-    {
-        if ($entity::class === 'App\Entity\Region'
-            || $entity::class === 'App\Entity\Department'
-            || $entity::class === 'Proxies\__CG__\App\Entity\Region'
-            || $entity::class === 'Proxies\__CG__\App\Entity\Department')
-        {
-            return $this->getVersionsAcrossPlaces($entity);
-        }
-        else
-        {
-            return $this->getVersionsDirectly($entity);
-        }
-    }
-
-    /** Get related versions for Region or Department.
-     * @param Region|Department $entity
-     *
-     * @return Version[]
-     */
-    public function getVersionsAcrossPlaces(Region|Department $entity): array
-    {
-        $places = $this->getPlacesByEntity($entity);
-        $versions = array();
-
-        foreach ($places as $place) {
-            $versions = array_merge($versions, $this->getVersionsByPlace($place));
-        }
-
-        return $versions;
-    }
-
-    /** Get related versions for Source or Type.
-     * @param Type|Source $entity
-     *
-     * @return array<int, Version>
-     */
-    public function getVersionsDirectly(Type|Source $entity): array
-    {
-        $filteredVersions = [];
-
-        foreach ($this->versions as $version) {
-            if ($entity::class === 'App\Entity\Type'
-                && $version->getType() !== null
-                && $version->getType()->getId() === $entity->getId())
-            {
-                $filteredVersions[] = $version;
-            }
-            elseif ($entity::class === 'App\Entity\Source'
-                && $version->getSource() !== null
-                && $version->getSource()->getId() === $entity->getId())
-            {
-                $filteredVersions[] = $version;
-            }
-        }
-
-        return $filteredVersions;
-    }
-
-    /** Get related places for Region or Department.
-     * @param Region|Department $entity
-     *
-     * @return array<int, Place>
-     */
-    public function getPlacesByEntity(Region|Department $entity): array
-    {
-        $filteredPlaces = [];
-
-        foreach ($this->places as $place) {
-            if ($entity::class === 'App\Entity\Region'
-                && $place->getRegion() !== null
-                && $place->getRegion()->getId() === $entity->getId())
-            {
-                $filteredPlaces[] = $place;
-            }
-            elseif ($entity::class === 'App\Entity\Department' // TODO: Remove association between Place & Department
-                && $place->getDepartment() !== null
-                && $place->getDepartment()->getId() === $entity->getId())
-            {
-                $filteredPlaces[] = $place;
-            }
-        }
-
-        return $filteredPlaces;
-    }
-
-    /** Get related versions for Place.
-     * @param Place $place
-     *
-     * @return array<int, Version>
-     */
-    public function getVersionsByPlace(Place $place): array
-    {
-        $placeId = $place->getId();
-        $versions = array();
-
-        foreach ($this->versions as $version) {
-            if ($version->getPlace() !== null
-                && $version->getPlace()->getId() === $placeId)
-            {
-                $versions[] = $version;
-            }
-        }
-
-        return $versions;
     }
 
     /**
